Fix some warnings introduced in gcc-4.3. Fix raw_vector.
git-svn-id: https://spacegameengine.svn.sourceforge.net/svnroot/spacegameengine@510 469afb1b-a615-0410-ae09-9c7fe5ec7362

+++ src/src/x_window.cpp
@@ -199,7 +199,7 @@ void sge::window::dispatch()
 int handler(Display* const d, XErrorEvent* const e)
 {
 	boost::array<char,1024> buf;
-	XGetErrorText(d, e->error_code,buf.c_array(), buf.size());
+	XGetErrorText(d, e->error_code,buf.c_array(), static_cast<int>(buf.size()));
 	std::cerr << "X Error: " << buf.data() << '\n';
 	return 0;
 }
